Rename gcp to defaultHttpAdapter
