+++ packages/circus-cs-core/src/job/adapters/defaultHttpAdapter.ts
