+++ packages/circus-cs-core/src/job/adapters/index.ts
@@ -1,6 +1,6 @@
-import gcp from './gcp';
+import defaultHttpAdapter from './defaultHttpAdapter';
 import { RemoteAdapter } from '../../util/remote-job';
 
-const adapters: Record<string, RemoteAdapter> = { gcp };
+const adapters: Record<string, RemoteAdapter> = { defaultHttpAdapter };
 
 export default adapters;
